+++ pyqcrbox/pyqcrbox/sql_models/parameter_spec/filesystem_path_parameters.py
@@ -8,26 +8,39 @@ from .base_parameter_spec import BaseParameterSpec
 __all__ = [
     "FolderPathParameterSpec",
     "InputCifParameterSpec",
-    "GenericInputFileParameterSpec",
+    "GenericInputPathParameterSpec",
     "OutputCifParameterSpec",
     "WorkCifParameterSpec",
+    "DataFileParameterSpec",
 ]
 
 
 class BaseFilesystemPathParameterSpec(BaseParameterSpec):
-    pass
+    def dtype_is_compatible_with(self, other_dtype: str):
+        if super().dtype_is_compatible_with(other_dtype):
+            return True
+
+        return other_dtype in ["str"]
 
 
-class GenericInputFileParameterSpec(BaseFilesystemPathParameterSpec):
-    dtype: Literal["QCrBox.input_file"]
+class GenericInputPathParameterSpec(BaseFilesystemPathParameterSpec):
+    dtype: Literal["QCrBox.input_path"]
 
 
-class GenericOutputFileParameterSpec(BaseFilesystemPathParameterSpec):
-    dtype: Literal["QCrBox.output_file"]
+class GenericOutputPathParameterSpec(BaseFilesystemPathParameterSpec):
+    dtype: Literal["QCrBox.output_path"]
+
+
+class InputFolderParameterSpec(BaseFilesystemPathParameterSpec):
+    dtype: Literal["QCrBox.input_folder"]
 
 
 class FolderPathParameterSpec(BaseFilesystemPathParameterSpec):
     dtype: Literal["QCrBox.folder_path"]
+
+
+class DataFileParameterSpec(BaseFilesystemPathParameterSpec):
+    dtype: Literal["QCrBox.data_file"]
 
 
 class BaseCifFileParameterSpec(BaseFilesystemPathParameterSpec):