+++ services/applications/mopro/configure_mopro.py
@@ -1,14 +1,75 @@
 import os
 import re
+import shlex
+
+# import os
+import subprocess
 from pathlib import Path, PureWindowsPath
 
-from qcrboxtools.cif.cif2cif import cif_file_merge_to_unified_by_yml, cif_file_to_specific_by_yml
+from qcrboxtools.cif.cif2cif import cif_file_merge_to_unified_by_yml, cif_file_to_specific_by_yml, cif_file_to_unified
 from qcrboxtools.cif.file_converter.hkl import cif2hkl4
 
 from pyqcrbox import sql_models
 from pyqcrbox.registry.client import QCrBoxClient
 
 YAML_PATH = "./config_mopro.yaml"
+
+
+def __run_interactive(input_file):
+    input_cif_path = Path(input_file)
+    work_dir = input_cif_path.parent
+    roaming_dir = Path("/opt/wine_installations/wine_win64/drive_c/users/qcrbox/AppData/Roaming/mopro")
+    roaming_dir.mkdir(parents=True, exist_ok=True)
+    wine_work_dir = PureWindowsPath("Z:\\opt\\qcrbox")
+    # for part in Path(work_dir.absolute()).parts[4:]:
+    #    wine_work_dir = wine_work_dir / part
+
+    replace_dict = {"{{mopro_workdir}}": str(wine_work_dir)}
+    for filename in Path("./templates").iterdir():
+        with Path(filename).open("r", encoding="UTF-8") as fobj:
+            content = fobj.read()
+        for key, value in replace_dict.items():
+            content = content.replace(key, value)
+        with (roaming_dir / filename.name).open("w", encoding="UTF-8") as fobj:
+            fobj.write(content)
+
+    command = shlex.split(
+        "wine /opt/wine_installations/wine_win64/drive_c/MoProSuite-win-July2022/MoProGUI_Qt_win64/MoProGUI_win64.exe"
+    )
+    # TODO write hkl
+    unified_cif_path = work_dir / "unified_for_hkl.cif"
+    cif_file_to_unified(input_cif_path, unified_cif_path)
+    cif2hkl4(unified_cif_path, 0, "shelx.hkl")
+    unified_cif_path.unlink()
+
+    subprocess.call(command)
+    # os.spawnl(os.P_NOWAIT, command)
+
+
+def __finalise_interactive(input_file):
+    input_cif_path = Path(input_file)
+    work_folder = input_cif_path.parent
+    try:
+        excluded_cif = ("output.cif", "work.cif", "input.cif")
+        newest_cif_path = next(
+            reversed(
+                sorted(
+                    (file_path for file_path in work_folder.glob("*.cif") if file_path.name not in excluded_cif),
+                    key=os.path.getmtime,
+                )
+            )
+        )
+
+        # MoPro might output invalid characters
+
+        text = newest_cif_path.read_text(encoding="utf-8", errors="replace")
+        non_character_pattern = re.compile(r"[^\w\s\.,!?;:\'\"\-()\[\]{}<>|/\\@#%&*+=`~]")
+        cleaned_text = non_character_pattern.sub("?", text)
+        cleaned_cif_path = newest_cif_path.with_name("cleaned.cif")
+        cleaned_cif_path.write_text(cleaned_text, encoding="utf-8")
+        return cleaned_cif_path
+    except StopIteration:
+        pass
 
 
 def prepare__interactive(input_cif_path):
@@ -93,6 +154,7 @@ def finalise__interactive(input_cif_path, output_cif_path):
         )
     except StopIteration:
         pass
+
 
 if __name__ == "__main__":
     application_spec = sql_models.ApplicationSpec.from_yaml_file(YAML_PATH)